+++ package.json
@@ -27,6 +27,7 @@
     "grunt": "^0.4.5",
     "grunt-docco": "^0.4.0",
     "grunt-nodemon": "^0.4.0",
+    "mkpath": "^1.0.0",
     "morgan": "^1.6.1",
     "mysql": "^2.9.0",
     "passport": "^0.3.2",
