--- server/tmp/Awesome%20Snippet.sublime-snippet
@@ -1,7 +1,0 @@
-<snippet>
-<content><![CDATA[
-snippets are awesomer
-]]></content>
-<tabTrigger>hello</tabTrigger>
-<scope>scope.javascript</scope>
-</snippet>
