--- server/tmp/Awesome%20Snippet1.sublime-snippet
@@ -1,7 +1,0 @@
-<snippet>
-<content><![CDATA[
-snippets are awesome
-]]></content>
-<tabTrigger>hello</tabTrigger>
-<scope>scope.javascript</scope>
-</snippet>
