--- server/tmp/Awesome%20Snippet2.sublime-snippet
@@ -1,7 +1,0 @@
-<snippet>
-<content><![CDATA[
-snippets are awesomest
-]]></content>
-<tabTrigger>hello</tabTrigger>
-<scope>scope.javascript</scope>
-</snippet>
